% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneseng_read_blog.R
\name{geneseng_read_blog}
\alias{geneseng_read_blog}
\title{Knit a Geneseng article}
\usage{
geneseng_read_blog(path)
}
\arguments{
\item{path}{path of the file}
}
\description{
Knit a Geneseng article
}
